src/requirements.txt
grpcio==1.59.2
protobuf==4.25.1
pydantic==2.5.1
SQLAlchemy==2.0.23
grpcio-tools
python-dotenv==1.0.0
psycopg2==2.9.9
